ml/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
torch>=2.6.0
torchvision>=0.17.0
transformers>=4.36.2
ultralytics>=8.1.0
onnxruntime>=1.16.3
sentence-transformers>=2.2.2
Pillow>=10.1.0
numpy>=1.26.2
requests>=2.31.0
gunicorn>=21.2.0
